Added resource audit and logging
An audit of resources functionality has been added. The data will be
logged into an audit.txt.
The audit data is also stored in a hash struct where it will be used to
detect vCPU over commitment. Also, added checking functionality to the
min_create where formula is used to ensure there are enough vCPUs
available to create x number of instances by the daemon.